Conglomerate each network's ping_bootnodes methods into single overlay method
