+++ trin-state/src/network.rs
@@ -2,7 +2,6 @@ use std::sync::Arc;
 
 use discv5::enr::NodeId;
 use eth_trie::EthTrie;
-use log::{debug, error};
 use parking_lot::RwLock;
 use tokio::sync::mpsc::UnboundedSender;
 use trin_core::{
@@ -64,36 +63,6 @@ impl StateNetwork {
         Self {
             overlay: Arc::new(overlay),
             trie: Arc::new(trie),
-        }
-    }
-
-    /// Convenience call for testing, quick way to ping bootnodes
-    pub async fn ping_bootnodes(&self) {
-        // Trigger bonding with bootnodes, at both the base layer and portal overlay.
-        // The overlay ping via talkreq will trigger a session at the base layer, then
-        // a session on the (overlay) portal network.
-        let mut successfully_bonded_bootnode = false;
-        let enrs = self.overlay.discovery.discv5.table_entries_enr();
-        if enrs.is_empty() {
-            error!("No bootnodes provided, cannot join Portal State Network.");
-            return;
-        }
-        for enr in enrs {
-            debug!("Attempting bond with bootnode {}", enr);
-            let ping_result = self.overlay.send_ping(enr.clone()).await;
-
-            match ping_result {
-                Ok(_) => {
-                    debug!("Successfully bonded with {}", enr);
-                    successfully_bonded_bootnode = true;
-                }
-                Err(err) => {
-                    error!("{err} while pinging bootnode: {enr:?}");
-                }
-            }
-        }
-        if !successfully_bonded_bootnode {
-            error!("Failed to bond with any bootnodes, cannot join Portal State Network.");
         }
     }
 }
